requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
scikit-learn
sqlmodel
asyncpg
passlib[bcrypt]
python-dotenv